requirements.txt
pandas==1.5.2
tqdm==4.64.0
seaborn==0.11.2
statsmodels==0.13.2
jupyter==1.0.0
plotly==5.13.1
protobuf>=3.12,<4
google-cloud-bigquery==3.7.0
pandas-gbq==0.19.1
streamlit==1.20.0